+++ src/main/java/com/example/DBkursova/entity/Equipment.java
@@ -18,7 +18,7 @@ public class Equipment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "equipment_id")
-    private int idEquipment;
+    private Long idEquipment;
     @Column(name = "equipment_type")
     private String type;
     @Column(name = "equipment_name")
@@ -26,7 +26,7 @@ public class Equipment {
     @Column(name = "equipment_condition")
     private String condition;
     @Column(name = "equipment_cost")
-    private int cost;
+    private Integer cost;
     @ManyToMany(mappedBy = "equipment", cascade = CascadeType.ALL)
     private List<RoadObject> object = new ArrayList<>();
 }
